buildPortRange method bug-fix (and improved)
---

Bug-fix: to be compliant with the OpenWRT documentation, it is necessary to specify the port range with a dash instead of using the column.

More info on [OpenWRT doc](https://openwrt.org/docs/guide-user/firewall/firewall_configuration#rules)

---

If the port range is not specified the MUD manager will not work properly.
For this reason, I updated buildPortRange function to replace NULL with "any".

---